services/event_to_notification/requirements.txt
backoff==2.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
confluent-kafka==1.8.2
idna==3.3
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
orjson==3.6.9
pydantic==1.9.1
PyYAML==6.0
requests==2.27.1
typing_extensions==4.2.0
urllib3==1.26.9
